plugins/audiobook-plugin/skills/audiobook/scripts/pyproject.toml
[project]
name = "audiobook-tts"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.61.0",
    "mlx-audio==0.3.0",
    "tqdm>=4.67.2",
]
